refactor(core): extract ffi test owner
